[Respects] Handle case where bot has no Send Message permissions (#214)
- Resolves #213.
- Keep tabs on the user that has paid respects.
- Do a check to see if the discord.Message object exists first before trying to access its attributes.

This fixes the case where someone types [p]f in a channel where send permissions are denied (e.g. #dark-hour). When this happens, the message object is not saved to ...[KEY_MSG] as it errors out in line 228. At that point, the sid and cid of the server and channel IDs have been set in the dict, respectively. This fixes the case where if send message perms are re-enabled in the channel, the check above doesn't error out due to accessing a non-existent Message object, since we already had a previous respect being paid (albeit didn't send properly).