[VxTwit] Add per-guild toggle
This commit adds the ability to toggle the VxTwitter replacements on a
per-guild basis. It requires the use of Config to store the enabled
state, which is then used within the event handler.

By default, it is disabled. It can be toggled by the guild admin using the
`[p]vxtwit toggle` command.