[SFU] Handle ConnectionError exception instead of MaxRetryError
When we introduced retries in SFUAnime/Ren#591, we expected to handle
the MaxRetryError exception. Instead, what appears to happen is the
MaxRetryError exception block isn't hit, and instead, we get a chain
of Exceptions that results in the underlying Exception being raised,
which is a ConnectionError.

As such, this commit changes the MaxRetryError block to a
ConnectionError block, thus returning a more descriptive message to the
end user. An early return is also added to avoid a UnboundLocalError
when trying to use fetchedData after the Exception was handled.

This addresses SFUAnime/Ren#663.